+++ src/ldc/supervised/classification/_csv.py
@@ -250,7 +250,7 @@ class AbstractCsvLikeClassificationWriter(BatchClassificationWriter, abc.ABC, Pl
 
     def __init__(self, target: str = None, no_header: bool = False,
                  col_text: str = None, col_label: str = None, col_id: str = None,
-                 logger_name: str = None, logging_level: str = LOGGING_WARNING):
+                 encoding: str = None, logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
         Initializes the writer.
 
@@ -264,6 +264,8 @@ class AbstractCsvLikeClassificationWriter(BatchClassificationWriter, abc.ABC, Pl
         :type col_label: str
         :param col_id: the (optional) column containing row IDs
         :type col_id: str
+        :param encoding: the encoding to use, None for default
+        :type encoding: str
         :param logger_name: the name to use for the logger
         :type logger_name: str
         :param logging_level: the logging level to use
@@ -275,6 +277,7 @@ class AbstractCsvLikeClassificationWriter(BatchClassificationWriter, abc.ABC, Pl
         self.col_text = col_text
         self.col_label = col_label
         self.col_id = col_id
+        self.encoding = encoding
         self._current_output = None
         self._output = None
         self._output_writer = None
@@ -301,6 +304,7 @@ class AbstractCsvLikeClassificationWriter(BatchClassificationWriter, abc.ABC, Pl
         parser.add_argument("--col_label", metavar="COL", type=str, default=None, help="The name of the column for the labels", required=False)
         parser.add_argument("--col_id", metavar="COL", type=str, default=None, help="The name of the column for the row IDs (uses 'id' from meta-data)", required=False)
         parser.add_argument("-n", "--no_header", action="store_true", help="For suppressing the header row", required=False)
+        parser.add_argument("--encoding", metavar="ENC", type=str, default=None, help="The encoding to force instead of using the default, e.g., 'utf-8'", required=False)
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -316,6 +320,7 @@ class AbstractCsvLikeClassificationWriter(BatchClassificationWriter, abc.ABC, Pl
         self.col_label = ns.col_label
         self.col_id = ns.col_id
         self.no_header = ns.no_header
+        self.encoding = ns.encoding
 
     def initialize(self):
         """
@@ -356,7 +361,7 @@ class AbstractCsvLikeClassificationWriter(BatchClassificationWriter, abc.ABC, Pl
             self.finalize()
             self._current_output = generate_output(self.session.current_input, target, self._get_extension(), self.session.options.compression)
             self.logger().info("Writing to: " + self._current_output)
-            self._output = open_file(self._current_output, mode="wt")
+            self._output = open_file(self._current_output, mode="wt", encoding=self.encoding)
             self._output_writer = self._init_writer(self._output)
             if not self.no_header:
                 row = []