+++ src/ldc/supervised/pairs/_csv.py
@@ -263,7 +263,7 @@ class AbstractCsvLikePairsWriter(BatchPairWriter, abc.ABC, PlaceholderSupporter)
 
     def __init__(self, target: str = None, no_header: bool = False,
                  col_instruction: str = None, col_input: str = None, col_output: str = None, col_id: str = None,
-                 logger_name: str = None, logging_level: str = LOGGING_WARNING):
+                 encoding: str = None, logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
         Initializes the writer.
 
@@ -279,6 +279,8 @@ class AbstractCsvLikePairsWriter(BatchPairWriter, abc.ABC, PlaceholderSupporter)
         :type col_output: str
         :param col_id: the (optional) column containing row IDs
         :type col_id: str
+        :param encoding: the encoding to use, None for default
+        :type encoding: str
         :param logger_name: the name to use for the logger
         :type logger_name: str
         :param logging_level: the logging level to use
@@ -291,6 +293,7 @@ class AbstractCsvLikePairsWriter(BatchPairWriter, abc.ABC, PlaceholderSupporter)
         self.col_input = col_input
         self.col_output = col_output
         self.col_id = col_id
+        self.encoding = encoding
         self._current_output = None
         self._output = None
         self._output_writer = None
@@ -318,6 +321,7 @@ class AbstractCsvLikePairsWriter(BatchPairWriter, abc.ABC, PlaceholderSupporter)
         parser.add_argument("--col_output", metavar="COL", type=str, default=None, help="The name of the column for the outputs", required=False)
         parser.add_argument("--col_id", metavar="COL", type=str, default=None, help="The name of the column for the row IDs (uses 'id' from meta-data)", required=False)
         parser.add_argument("-n", "--no_header", action="store_true", help="For suppressing the header row", required=False)
+        parser.add_argument("--encoding", metavar="ENC", type=str, default=None, help="The encoding to force instead of using the default, e.g., 'utf-8'", required=False)
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -334,6 +338,7 @@ class AbstractCsvLikePairsWriter(BatchPairWriter, abc.ABC, PlaceholderSupporter)
         self.col_output = ns.col_output
         self.col_id = ns.col_id
         self.no_header = ns.no_header
+        self.encoding = ns.encoding
 
     def initialize(self):
         """
@@ -374,7 +379,7 @@ class AbstractCsvLikePairsWriter(BatchPairWriter, abc.ABC, PlaceholderSupporter)
             self.finalize()
             self._current_output = generate_output(self.session.current_input, target, self._get_extension(), self.session.options.compression)
             self.logger().info("Writing to: " + self._current_output)
-            self._output = open_file(self._current_output, mode="wt")
+            self._output = open_file(self._current_output, mode="wt", encoding=self.encoding)
             self._output_writer = self._init_writer(self._output)
             if not self.no_header:
                 row = []